unified integration of all features
