+++ package.json
@@ -11,10 +11,8 @@
   "dependencies": {
     "next": "14.2.15",
     "react": "^18",
-
     "react-dom": "^18",
     "react-icons": "^5.3.0"
-
   },
   "devDependencies": {
     "autoprefixer": "^10.4.20",
